refactor(runtime): separate contexts limits events and lifecycle
